refactor(vector): move embedder to vector extension
