+++ src/core/embedder.ts
@@ -1,167 +1,4 @@
-/**
- * Local Embedding Generator using @xenova/transformers
- * AXIOMMIND Principle 7: Standard JSON format for vectors
- */
-
-export interface EmbeddingResult {
-  vector: number[];
-  model: string;
-  dimensions: number;
-}
-
-type FeatureExtractionPipelineFactory = (
-  task: 'feature-extraction',
-  model: string
-) => Promise<NonNullable<Embedder['pipeline']>>;
-
-export class Embedder {
-  private pipeline: ((input: string, options?: Record<string, unknown>) => Promise<{ data: Float32Array }>) | null = null;
-  private readonly modelName: string;
-  private activeModelName: string;
-  private initialized = false;
-
-  constructor(modelName: string = 'jinaai/jina-embeddings-v5-text-nano-text-matching') {
-    this.modelName = modelName;
-    this.activeModelName = modelName;
-  }
-
-  /**
-   * Initialize the embedding pipeline
-   */
-  async initialize(): Promise<void> {
-    if (this.initialized) return;
-
-    const pipeline = await loadTransformersPipeline();
-
-    try {
-      this.pipeline = await pipeline('feature-extraction', this.modelName);
-      this.activeModelName = this.modelName;
-      this.initialized = true;
-      return;
-    } catch (primaryError) {
-      const fallbackModel = process.env.CLAUDE_MEMORY_EMBEDDING_FALLBACK_MODEL || 'onnx-community/embeddinggemma-300m-ONNX';
-      if (fallbackModel === this.modelName) {
-        throw primaryError;
-      }
-
-      console.warn(`[Embedder] Primary model failed (${this.modelName}). Falling back to ${fallbackModel}`);
-      this.pipeline = await pipeline('feature-extraction', fallbackModel);
-      this.activeModelName = fallbackModel;
-      this.initialized = true;
-    }
-  }
-
-  // ~4 chars per token; 512 tokens * 4 = 2048, use 2000 to be safe
-  private static readonly MAX_CHARS = 2000;
-
-  private truncate(text: string): string {
-    return text.length > Embedder.MAX_CHARS ? text.slice(0, Embedder.MAX_CHARS) : text;
-  }
-
-  /**
-   * Generate embedding for a single text
-   */
-  async embed(text: string): Promise<EmbeddingResult> {
-    await this.initialize();
-
-    if (!this.pipeline) {
-      throw new Error('Embedding pipeline not initialized');
-    }
-
-    const output = await this.pipeline(this.truncate(text), {
-      pooling: 'mean',
-      normalize: true,
-      truncation: true,
-      max_length: 512
-    });
-
-    const vector = Array.from(output.data);
-
-    return {
-      vector,
-      model: this.activeModelName,
-      dimensions: vector.length
-    };
-  }
-
-  /**
-   * Generate embeddings for multiple texts in batch
-   */
-  async embedBatch(texts: string[]): Promise<EmbeddingResult[]> {
-    await this.initialize();
-
-    if (!this.pipeline) {
-      throw new Error('Embedding pipeline not initialized');
-    }
-
-    const results: EmbeddingResult[] = [];
-
-    // Process in batches of 32 for memory efficiency
-    const batchSize = 32;
-    for (let i = 0; i < texts.length; i += batchSize) {
-      const batch = texts.slice(i, i + batchSize);
-
-      for (const text of batch) {
-        const output = await this.pipeline(this.truncate(text), {
-          pooling: 'mean',
-          normalize: true,
-          truncation: true,
-          max_length: 512
-        });
-
-        const vector = Array.from(output.data);
-
-        results.push({
-          vector,
-          model: this.activeModelName,
-          dimensions: vector.length
-        });
-      }
-    }
-
-    return results;
-  }
-
-  /**
-   * Get embedding dimensions for the current model
-   */
-  async getDimensions(): Promise<number> {
-    const result = await this.embed('test');
-    return result.dimensions;
-  }
-
-  /**
-   * Check if embedder is ready
-   */
-  isReady(): boolean {
-    return this.initialized && this.pipeline !== null;
-  }
-
-  /**
-   * Get model name
-   */
-  getModelName(): string {
-    return this.activeModelName;
-  }
-}
-
-// Singleton instance for reuse
-let defaultEmbedder: Embedder | null = null;
-
-export function getDefaultEmbedder(): Embedder {
-  const envModel = process.env.CLAUDE_MEMORY_EMBEDDING_MODEL;
-  if (!defaultEmbedder) {
-    defaultEmbedder = new Embedder(envModel || undefined);
-  }
-  return defaultEmbedder;
-}
-
-async function loadTransformersPipeline(): Promise<FeatureExtractionPipelineFactory> {
-  // Keep @huggingface/transformers lazy so importing MemoryService or pure
-  // adapter helpers does not eagerly dlopen onnxruntime native bindings.
-  const dynamicImport = new Function('specifier', 'return import(specifier)') as (
-    specifier: string
-  ) => Promise<{ pipeline: unknown }>;
-  const transformers = await dynamicImport('@huggingface/transformers');
-  return transformers.pipeline as FeatureExtractionPipelineFactory;
-}
+// Compatibility re-export. Embedding is an optional vector extension, but
+// existing core/engine imports still resolve through this path during the
+// strangler migration.
+export * from '../extensions/vector/index.js';
